refactor(role-archive): consolidate role archive import/export pipeline
将 8 个连续 commit 合并为单个逻辑单元，覆盖角色压缩包导入/导出
功能的前后端重构 + UI 调整 + 文档更新。

UI — src/components/ui/ImportProgressBar.vue
- inline tailwind class，把青色 token 统一到 #79d9ff
- 修复 mojibake 中文
- 把动态 Tailwind class 拼接换成 scoped CSS + [data-phase=...] 静态选择器
- archive-shimmer 关键帧去 scoped 化以匹配动态 class

后端 — src-tauri/src/utils/archive
- 拆分单文件 archive.rs (984 行) 为 mod/safety/resolve/extract/compress 子模块
- 移除 src-tauri/tests/archive/ 集成测试套件 (17 个测试)
- 给公开 API 加 rustdoc
- 保留所有 15 个核心函数语义（逐函数 diff 验证）

后端 — src-tauri/src/api/role_archive
- sanitize_role_folder_name 只保留字母数字 + - . 字符
- cancel_role_import 接受 task_id 参数
- 重复的 clearTimers() 调用去重

文档 — docs/
- 删除过期的 docs/utils/archive.md（路径已变 + 正文乱码）
- 新增 docs/utils/role-archive.md（中文），描述导入/导出端到端流程

文件: 20 changed, 1276 insertions(+), 1188 deletions(-)
